working through refactoring
more refactoring

major refactor
