+++ cmd/kops/delete_instancegroup.go
@@ -18,7 +18,6 @@ package main
 
 import (
 	"fmt"
-
 	"io"
 	"os"
